refactor: ファイル/フォルダ名の命名規則として接続部分を-に統一
