+++ src/three/cameras/front-camera.ts
