+++ src/three/cameras/main-camera.ts
