+++ src/three/cameras/side-camera.ts
