+++ src/three/cameras/top-camera.ts
