+++ src/three/images/annotated-box.ts
