+++ src/three/images/camera-image.ts
